+++ src/frameworks/upstream/mod.rs
@@ -1,3 +1,5 @@
 pub mod dns_client;
+pub mod dot_client;
 
 pub use dns_client::DnsUdpTcpClient;
+pub use dot_client::DnsTlsClient;